Send image based on video cap size. Send image continously
