+++ bots/web_bot_adapter/web_bot_adapter.py
@@ -13,6 +13,7 @@ import requests
 import undetected_chromedriver as uc
 from pyvirtualdisplay import Display
 from websockets.sync.server import serve
+from bots.utils import half_ceil, scale_i420
 
 from bots.bot_adapter import BotAdapter
 from bots.bot_controller.automatic_leave_configuration import AutomaticLeaveConfiguration
@@ -20,116 +21,6 @@ from bots.bot_controller.automatic_leave_configuration import AutomaticLeaveConf
 from .ui_methods import UiRequestToJoinDeniedException, UiRetryableException
 
 logger = logging.getLogger(__name__)
-
-
-def half_ceil(x):
-    return (x + 1) // 2
-
-
-def scale_i420(frame, frame_size, new_size):
-    """
-    Scales an I420 (YUV 4:2:0) frame from 'frame_size' to 'new_size',
-    handling odd frame widths/heights by using 'ceil' in the chroma planes.
-
-    :param frame:      A bytes object containing the raw I420 frame data.
-    :param frame_size: (orig_width, orig_height)
-    :param new_size:   (new_width, new_height)
-    :return:           A bytes object with the scaled I420 frame.
-    """
-
-    # 1) Unpack source / destination dimensions
-    orig_width, orig_height = frame_size
-    new_width, new_height = new_size
-
-    # 2) Compute source plane sizes with rounding up for chroma
-    orig_chroma_width = half_ceil(orig_width)
-    orig_chroma_height = half_ceil(orig_height)
-
-    y_plane_size = orig_width * orig_height
-    uv_plane_size = orig_chroma_width * orig_chroma_height  # for each U or V
-
-    # 3) Extract Y, U, V planes from the byte array
-    y = np.frombuffer(frame[0:y_plane_size], dtype=np.uint8)
-    u = np.frombuffer(frame[y_plane_size : y_plane_size + uv_plane_size], dtype=np.uint8)
-    v = np.frombuffer(
-        frame[y_plane_size + uv_plane_size : y_plane_size + 2 * uv_plane_size],
-        dtype=np.uint8,
-    )
-
-    # 4) Reshape planes
-    y = y.reshape(orig_height, orig_width)
-    u = u.reshape(orig_chroma_height, orig_chroma_width)
-    v = v.reshape(orig_chroma_height, orig_chroma_width)
-
-    # ---------------------------------------------------------
-    # Scale preserving aspect ratio or do letterbox/pillarbox
-    # ---------------------------------------------------------
-    input_aspect = orig_width / orig_height
-    output_aspect = new_width / new_height
-
-    if abs(input_aspect - output_aspect) < 1e-6:
-        # Same aspect ratio; do a straightforward resize
-        scaled_y = cv2.resize(y, (new_width, new_height), interpolation=cv2.INTER_LINEAR)
-
-        # For U, V we should scale to half-dimensions (rounded up)
-        # of the new size. But OpenCV requires exact (int) dims, so:
-        target_u_width = half_ceil(new_width)
-        target_u_height = half_ceil(new_height)
-
-        scaled_u = cv2.resize(u, (target_u_width, target_u_height), interpolation=cv2.INTER_LINEAR)
-        scaled_v = cv2.resize(v, (target_u_width, target_u_height), interpolation=cv2.INTER_LINEAR)
-
-        # Flatten and return
-        return np.concatenate([scaled_y.flatten(), scaled_u.flatten(), scaled_v.flatten()]).astype(np.uint8).tobytes()
-
-    # Otherwise, the aspect ratios differ => letterbox or pillarbox
-    if input_aspect > output_aspect:
-        # The image is relatively wider => match width, shrink height
-        scaled_width = new_width
-        scaled_height = int(round(new_width / input_aspect))
-    else:
-        # The image is relatively taller => match height, shrink width
-        scaled_height = new_height
-        scaled_width = int(round(new_height * input_aspect))
-
-    # 5) Resize Y, U, and V to the scaled dimensions
-    scaled_y = cv2.resize(y, (scaled_width, scaled_height), interpolation=cv2.INTER_LINEAR)
-
-    # For U, V, use half-dimensions of the scaled result, rounding up.
-    scaled_u_width = half_ceil(scaled_width)
-    scaled_u_height = half_ceil(scaled_height)
-    scaled_u = cv2.resize(u, (scaled_u_width, scaled_u_height), interpolation=cv2.INTER_LINEAR)
-    scaled_v = cv2.resize(v, (scaled_u_width, scaled_u_height), interpolation=cv2.INTER_LINEAR)
-
-    # 6) Create the output buffers. For "dark" black:
-    #    Y=0, U=128, V=128.
-    final_y = np.zeros((new_height, new_width), dtype=np.uint8)
-    final_u = np.full((half_ceil(new_height), half_ceil(new_width)), 128, dtype=np.uint8)
-    final_v = np.full((half_ceil(new_height), half_ceil(new_width)), 128, dtype=np.uint8)
-
-    # 7) Compute centering offsets for each plane (Y first)
-    offset_y = (new_height - scaled_height) // 2
-    offset_x = (new_width - scaled_width) // 2
-
-    final_y[offset_y : offset_y + scaled_height, offset_x : offset_x + scaled_width] = scaled_y
-
-    # Offsets for U and V planes are half of the Y offsets (integer floor)
-    offset_y_uv = offset_y // 2
-    offset_x_uv = offset_x // 2
-
-    final_u[
-        offset_y_uv : offset_y_uv + scaled_u_height,
-        offset_x_uv : offset_x_uv + scaled_u_width,
-    ] = scaled_u
-    final_v[
-        offset_y_uv : offset_y_uv + scaled_u_height,
-        offset_x_uv : offset_x_uv + scaled_u_width,
-    ] = scaled_v
-
-    # 8) Flatten back to I420 layout and return bytes
-
-    return np.concatenate([final_y.flatten(), final_u.flatten(), final_v.flatten()]).astype(np.uint8).tobytes()
-
 
 class WebBotAdapter(BotAdapter):
     def __init__(
@@ -566,3 +457,6 @@ class WebBotAdapter(BotAdapter):
 
     def send_raw_audio(self, bytes, sample_rate):
         logger.info("send_raw_audio not supported in google meet bots")
+
+    def send_raw_image(self, image_bytes):
+        logger.info("send_raw_image not supported in google meet bots")
